Detect Content-Type in ServeContent before compression.
We want to detect the Content-Type eagerly, before passing it off
to http.ServeContent. It's because http.ServeContent won't be able
to easily detect the original content type after content has been
gzip compressed.

We do this even for the last case that serves uncompressed data
so that it doesn't have to do duplicate work.

Add a small test for this.

Fixes #1.