check if Content-Encoding header is already set before compressing
In case the handler has already dealt with compression and applied it,
ServeContent shouldn't try to do duplicate work. Detect that situation
and serve content as is if so.

This also makes it possible for caller to indicate a file isn't worth
compressing by explicitly providing a Content-Encoding header:

	w.Header().Set("Content-Type", "image/png")
	w.Header()["Content-Encoding"] = nil
	httpgzip.ServeContent([...])

This is the other half of commit b1c53ac65af9fd3d354d6f9ad30a0cca35e173ed.

Move ServeContent tests to gzip_test.go file and rename tests to be
more idiomatic. Underscores are used in some example names, but they
shouldn't be used in test names.